Cleanup search list, move shared search faq to top level components
