+++ js/components/SearchFaq.jsx
