+++ js/entrypage/components/Splash.jsx
